requirements.txt

# requirements.txt

# Data manipulation and analysis
pandas==1.3.5
numpy==1.21.5

# Machine Learning
scikit-learn==1.0.2
xgboost==1.5.0

# Data fetching
yfinance==0.1.70

# Web framework
Flask==2.0.2

# Database
SQLAlchemy==1.4.31
psycopg2-binary==2.9.3  # For PostgreSQL support

# Task scheduling
APScheduler==3.8.1

# Model serialization
joblib==1.1.0

# Testing
pytest==7.0.1

# Jupyter (for notebooks)
jupyter==1.0.0
notebook==6.4.8

# Visualization (if needed)
matplotlib==3.5.1
seaborn==0.11.2

# API documentation (optional, but recommended)
flasgger==0.9.5

# Environment variable management
python-dotenv==0.19.2

# YAML support (for configuration files)
PyYAML==6.0

# Date and time handling
python-dateutil==2.8.2

# Progress bars (optional, but useful for long-running tasks)
tqdm==4.62.3